+++ tests/Backend/Feature/api/v1/ApplicationTest.php
@@ -37,6 +37,7 @@ class ApplicationTest extends TestCase
         $this->generalSettings->legal_notice_url = 'http://localhost';
         $this->generalSettings->privacy_policy_url = 'http://localhost';
         $this->generalSettings->toast_lifetime = 10;
+        $this->generalSettings->no_welcome_page = false;
         $this->generalSettings->save();
 
         $this->themeSettings->logo = 'testlogo.svg';
@@ -95,6 +96,7 @@ class ApplicationTest extends TestCase
                         'help_url' => 'http://localhost',
                         'legal_notice_url' => 'http://localhost',
                         'privacy_policy_url' => 'http://localhost',
+                        'no_welcome_page' => false,
                     ],
                     'theme' => [
                         'logo' => 'testlogo.svg',
@@ -146,6 +148,7 @@ class ApplicationTest extends TestCase
         $this->generalSettings->help_url = null;
         $this->generalSettings->legal_notice_url = null;
         $this->generalSettings->privacy_policy_url = null;
+        $this->generalSettings->no_welcome_page = true;
         $this->generalSettings->save();
 
         $this->bannerSettings->enabled = false;
@@ -175,6 +178,7 @@ class ApplicationTest extends TestCase
                         'help_url' => null,
                         'legal_notice_url' => null,
                         'privacy_policy_url' => null,
+                        'no_welcome_page' => true,
                     ],
                     'theme' => [
                         'logo' => 'testlogo.svg',