+++ tests/Backend/Feature/api/v1/SettingsTest.php
@@ -48,6 +48,7 @@ class SettingsTest extends TestCase
         $this->generalSettings->help_url = 'http://localhost/help';
         $this->generalSettings->legal_notice_url = 'http://localhost/legal';
         $this->generalSettings->privacy_policy_url = 'http://localhost/privacy';
+        $this->generalSettings->no_welcome_page = false;
         $this->generalSettings->save();
 
         $this->themeSettings->logo = 'testlogo.svg';
@@ -116,6 +117,7 @@ class SettingsTest extends TestCase
                     'general_help_url' => 'http://localhost/help',
                     'general_legal_notice_url' => 'http://localhost/legal',
                     'general_privacy_policy_url' => 'http://localhost/privacy',
+                    'general_no_welcome_page' => false,
 
                     'theme_logo' => 'testlogo.svg',
                     'theme_logo_dark' => 'testlogo-dark.svg',
@@ -175,6 +177,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => true,
 
             'theme_logo' => 'testlogo.svg',
             'theme_logo_dark' => 'testlogo-dark.svg',
@@ -235,6 +238,7 @@ class SettingsTest extends TestCase
                     'general_help_url' => 'http://localhost',
                     'general_legal_notice_url' => 'http://localhost',
                     'general_privacy_policy_url' => 'http://localhost',
+                    'general_no_welcome_page' => true,
 
                     'theme_logo' => 'testlogo.svg',
                     'theme_logo_dark' => 'testlogo-dark.svg',
@@ -302,6 +306,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'theme_logo_file' => UploadedFile::fake()->image('logo.svg'),
             'theme_logo_dark_file' => UploadedFile::fake()->image('logo-dark.svg'),
@@ -372,6 +377,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'theme_logo' => '/storage/image/logo.svg',
             'theme_logo_file' => UploadedFile::fake()->image('logo.svg'),
@@ -450,6 +456,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 111,
             'general_legal_notice_url' => 222,
             'general_privacy_policy_url' => 333,
+            'general_no_welcome_page' => 'notbool',
 
             'theme_logo' => '',
             'theme_logo_file' => 'notimagefile',
@@ -500,6 +507,7 @@ class SettingsTest extends TestCase
                 'general_help_url',
                 'general_legal_notice_url',
                 'general_privacy_policy_url',
+                'general_no_welcome_page',
 
                 'theme_logo',
                 'theme_logo_file',
@@ -587,6 +595,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'banner_enabled' => 0,
             'banner_message' => 'Welcome to Test!',
@@ -640,6 +649,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'banner_enabled' => 0,
             'banner_message' => 'Welcome to Test!',
@@ -708,6 +718,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'theme_logo' => 'testlogo.svg',
             'theme_logo_dark' => 'testlogo-dark.svg',
@@ -817,6 +828,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'theme_logo' => 'testlogo.svg',
             'theme_logo_dark' => 'testlogo-dark.svg',
@@ -911,6 +923,7 @@ class SettingsTest extends TestCase
             'general_help_url' => 'http://localhost',
             'general_legal_notice_url' => 'http://localhost',
             'general_privacy_policy_url' => 'http://localhost',
+            'general_no_welcome_page' => false,
 
             'theme_logo' => 'testlogo.svg',
             'theme_logo_dark' => 'testlogo-dark.svg',